+++ examples/cfd/ahmed_body_mgn/train.py
@@ -62,6 +62,7 @@ class MGNTrainer:
             data_dir=to_absolute_path(cfg.data_dir),
             split="train",
             num_samples=cfg.num_training_samples,
+            num_workers=cfg.num_dataset_workers,
         )
 
         # instantiate validation dataset
@@ -71,6 +72,7 @@ class MGNTrainer:
             data_dir=to_absolute_path(cfg.data_dir),
             split="validation",
             num_samples=cfg.num_validation_samples,
+            num_workers=cfg.num_dataset_workers,
         )
 
         # instantiate dataloader
@@ -259,7 +261,7 @@ def main(cfg: DictConfig) -> None:
         # save checkpoint
         if dist.world_size > 1:
             torch.distributed.barrier()
-        if dist.rank == 0:
+        if dist.rank == 0 and (epoch + 1) % cfg.checkpoint_save_freq == 0:
             save_checkpoint(
                 to_absolute_path(cfg.ckpt_path),
                 models=trainer.model,